add the javalin base for the api, start to write the API markdown, add a default web site to test (will be change)
