--- src/main/java/ch/heig/data/Flag.java
@@ -1,4 +1,0 @@
-package ch.heig.data;
-
-public record Flag(String name, int priority) {
-}
