--- src/main/java/ch/heig/data/Player.java
@@ -1,9 +1,0 @@
-package ch.heig.data;
-
-public class Player {
-    public final String username;
-
-    public Player(String username){
-        this.username=username;
-    }
-}
