--- src/main/java/ch/heig/data/Split.java
@@ -1,7 +1,0 @@
-package ch.heig.data;
-
-/**
- * @param time time in milli
- */
-public record Split(int flagIndex, Flag flag, float time) {
-}
