+++ src/main/java/ch/heigvd/data/Flag.java
@@ -1,0 +1,4 @@
+package ch.heigvd.data;
+
+public record Flag(String name, int priority) {
+}
