+++ src/main/java/ch/heigvd/data/Player.java
@@ -1,0 +1,9 @@
+package ch.heigvd.data;
+
+public class Player {
+    public final String username;
+
+    public Player(String username){
+        this.username=username;
+    }
+}
