+++ src/main/java/ch/heigvd/data/Split.java
@@ -1,0 +1,7 @@
+package ch.heigvd.data;
+
+/**
+ * @param time time in milli
+ */
+public record Split(int flagIndex, Flag flag, float time) {
+}
