+++ src/main/java/ch/heigvd/web/Api.java
@@ -1,0 +1,11 @@
+package ch.heigvd.web;
+
+import io.javalin.Javalin;
+
+import javax.naming.Context;
+
+public class Api {
+    private static void getData(Context context){
+
+    }
+}
