+++ src/main/resources/public/style.css
@@ -1,0 +1,16 @@
+html {
+    height: 100%;
+}
+body { 
+    padding: 0; 
+    margin: 0 ;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+}
+#unity-canvas { 
+    background: #231F20; 
+    width: 100%; 
+    height: 100%;
+}
